requirements.txt
imageio==2.8.0
numpy==1.18.2
pillow==7.1.1
